requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.26.0
matplotlib>=3.7.0
pyyaml>=6.0
scipy>=1.10.0
pandas>=2.0.0
tensorboard>=2.13.0
tqdm>=4.65.0
